Fix relationship updater test
