+++ src/agents/core/base_agent.py
@@ -25,10 +25,10 @@ except Exception:  # pragma: no cover - fallback when chromadb missing
     ChromaVectorStoreManager = None  # type: ignore[misc, assignment]
 from src.agents.memory.weaviate_vector_store_manager import WeaviateVectorStoreManager
 from src.infra import config
+from src.infra.async_dspy_manager import AsyncDSPyManager
 from src.infra.config import get_config
 from src.infra.llm_client import get_ollama_client
 from src.interfaces.dashboard_backend import AgentMessage, message_sse_queue
-from src.shared.async_utils import AsyncDSPyManager
 from src.shared.memory_store import MemoryStore
 from src.shared.typing import SimulationMessage
