+++ tests/unit/utils/test_async_dspy_manager_shutdown.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.shared.async_utils import AsyncDSPyManager
+from src.infra.async_dspy_manager import AsyncDSPyManager
 
 
 @pytest.mark.unit
